pyproject.toml
[tool.poetry]
name = "calendar-redacter"
version = "0.1.0"
description = "Creates a copy of a calendar with event details redacted"
authors = ["Radian LLC <contact+calendar-redacter@radian.codes>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
vdirsyncer = "^0.19.3"
python-dotenv = "^1.0.1"
ics = "^0.7.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
